+++ src/variorum/Intel/Intel_06_3E.c
@@ -73,8 +73,8 @@ static struct ivybridge_3e_offsets msrs =
 
 int intel_cpu_fm_06_3e_get_power_limits(int long_ver)
 {
-    uint16_t socket;
-    uint16_t nsockets, ncores, nthreads;
+    int socket;
+    int nsockets, ncores, nthreads;
 #ifdef VARIORUM_WITH_INTEL_CPU
     variorum_get_topology(&nsockets, &ncores, &nthreads, P_INTEL_CPU_IDX);
 #endif
@@ -151,8 +151,8 @@ int intel_cpu_fm_06_3e_get_power_limits(int long_ver)
 
 int intel_cpu_fm_06_3e_cap_power_limits(int package_power_limit)
 {
-    uint16_t socket;
-    uint16_t nsockets, ncores, nthreads;
+    int socket;
+    int nsockets, ncores, nthreads;
 #ifdef VARIORUM_WITH_INTEL_CPU
     variorum_get_topology(&nsockets, &ncores, &nthreads, P_INTEL_CPU_IDX);
 #endif
@@ -390,7 +390,7 @@ int intel_cpu_fm_06_3e_enable_turbo(void)
     {
         printf("Running %s\n", __FUNCTION__);
     }
-    unsigned int turbo_mode_disable_bit = 38;
+    int turbo_mode_disable_bit = 38;
     set_turbo_on(msrs.ia32_misc_enable, turbo_mode_disable_bit);
 
     return 0;
@@ -404,7 +404,7 @@ int intel_cpu_fm_06_3e_disable_turbo(void)
         printf("Running %s\n", __FUNCTION__);
     }
 
-    unsigned int turbo_mode_disable_bit = 38;
+    int turbo_mode_disable_bit = 38;
     set_turbo_off(msrs.ia32_misc_enable, turbo_mode_disable_bit);
 
     return 0;
@@ -418,7 +418,7 @@ int intel_cpu_fm_06_3e_get_turbo_status(void)
         printf("Running %s\n", __FUNCTION__);
     }
 
-    unsigned int turbo_mode_disable_bit = 38;
+    int turbo_mode_disable_bit = 38;
     print_turbo_status(stdout, msrs.ia32_misc_enable, turbo_mode_disable_bit);
 
     return 0;
@@ -520,7 +520,7 @@ int intel_cpu_fm_06_3e_cap_best_effort_node_power_limit(int node_limit)
      * the floor of the value being taken. So while we will be off by 1W total,
      * we will guarantee that we stay under the specified cap. */
 
-    uint16_t nsockets, ncores, nthreads;
+    int nsockets, ncores, nthreads;
 #ifdef VARIORUM_WITH_INTEL_CPU
     variorum_get_topology(&nsockets, &ncores, &nthreads, P_INTEL_CPU_IDX);
 #endif
