+++ src/variorum/config_architecture.h
@@ -319,9 +319,9 @@ int variorum_exit(
 );
 
 void variorum_get_topology(
-    uint16_t *nsockets,
-    uint16_t *ncores,
-    uint16_t *nthreads,
+    int *nsockets,
+    int *ncores,
+    int *nthreads,
     int idx
 );
